tlog_witness: extract shared error and helpers into common module
Move TlogWitnessError, MAX_CONSISTENCY_PROOF_LINES,
CONTENT_TYPE_TLOG_SIZE, and the parse_proof_line helper out of
add_checkpoint.rs into a new private common.rs module. The helper
becomes pub(crate) so additional endpoint modules can reuse it; the
public types continue to be re-exported from the crate root.

Pure refactor; no behavior change. Prepares for a sibling
sign_subtree module that needs the same error type, proof-line cap,
and base64 hash-line parser.
